+++ compiler/commands.nim
@@ -702,7 +702,14 @@ proc processSwitch(switch, arg: string, pass: TCmdLinePass, info: TLineInfo;
     expectNoArg(switch, arg, pass, info)
     useNimNamespace = true
     defineSymbol("cppCompileToNamespace")
-    
+  of "evaltemplatelimit":
+    expectArg(switch, arg, pass, info)
+    defineSymbol("evalTemplateLimit", arg)
+    evalTemplateLimit = parseInt(arg)
+  of "evalmacrolimit":
+    expectArg(switch, arg, pass, info)
+    defineSymbol("evalMacroLimit", arg)
+    evalMacroLimit = parseInt(arg)
   else:
     if strutils.find(switch, '.') >= 0: options.setConfigVar(switch, arg)
     else: invalidCmdLineOption(pass, switch, info)