+++ compiler/options.nim
@@ -149,6 +149,8 @@ var
   newDestructors*: bool
   gDynlibOverrideAll*: bool
   useNimNamespace*: bool
+  evalTemplateLimit* = 100
+  evalMacroLimit* = 100
 
 type
   SymbolFilesOption* = enum